Get `squall` command working:
I finally figured out how to get the C extension talking to Squall, which means
I can actually start using this for basic applications. There is still a lot
that needs to be added, like a CLI, a way to automatically traverse files, and
a way to ignore files as well.