Add `--db` flag:
This flag specifies the database to use for verifying SQL queries. Each
connection is made separately, meaning that calls that depend on one another
will fail. This was an issue before, but made more apperent by the use of
`:memory:` for the databases. This will be fixed at a later time.